requirements.txt
pandas==2.1.2
scikit-learn==1.2.2
scikit-surprise==1.1.1
protobuf==3.2.0
